fix: #162 log 목록 불러오는 중 상태 초기화 문제 해결
- `useSuspenseQuery` 를 `useQuery` + `isFetching` 으로 변경
- page.tsx 리팩토링
