+++ app/user/logs/components/CalendarHeader.tsx
@@ -6,17 +6,46 @@ import Icon from "@/ds/components/atoms/icon/Icon"
 import { CalendarHeaderProps } from "@/app/user/logs/types"
 
 export const CalendarHeader = ({
+  calendarRef,
   calMonth,
-  onNext,
-  onPrev,
-  onToday,
+  setCalMonth,
+  setSelectedDate,
 }: CalendarHeaderProps) => {
+  const setNewMonthTitle = () => {
+    if (calendarRef.current) {
+      const date = calendarRef.current.getApi().getDate()
+      const newMonth = `${date.getFullYear()}.${(date.getMonth() + 1).toString().padStart(2, "0")}`
+      setCalMonth(newMonth)
+      setSelectedDate(null)
+    }
+  }
+
+  const handleNext = () => {
+    if (calendarRef.current) {
+      calendarRef.current.getApi().next()
+      setNewMonthTitle()
+    }
+  }
+
+  const handlePrev = () => {
+    if (calendarRef.current) {
+      calendarRef.current.getApi().prev()
+      setNewMonthTitle()
+    }
+  }
+
+  const handleToday = () => {
+    if (calendarRef.current) {
+      calendarRef.current.getApi().today()
+      setNewMonthTitle()
+    }
+  }
   return (
     <div className="calendar-title-box grid w-full grid-cols-5 gap-2">
       <div></div>
       <div className="col-span-3 flex items-center justify-center gap-2">
         <div className="prev-button">
-          <Button variant="ghost" size="xs" onClick={onPrev}>
+          <Button variant="ghost" size="xs" onClick={handlePrev}>
             <Icon name="leftArrow" size={24} />
           </Button>
         </div>
@@ -24,13 +53,13 @@ export const CalendarHeader = ({
           <H1>{calMonth}</H1>
         </div>
         <div className="next-button">
-          <Button variant="ghost" size="xs" onClick={onNext}>
+          <Button variant="ghost" size="xs" onClick={handleNext}>
             <Icon name="rightArrow" size={24} />
           </Button>
         </div>
       </div>
       <div className="today-button flex items-center justify-center">
-        <Button variant="ghost" size="xs" onClick={onToday}>
+        <Button variant="ghost" size="xs" onClick={handleToday}>
           <C1 variant="secondary">today</C1>
         </Button>
       </div>
