+++ hooks/useGetUserLogs.ts
@@ -1,5 +1,5 @@
 import { getUserLogs } from "@/services/user/logs/getUserLogs"
-import { useSuspenseQuery } from "@tanstack/react-query"
+import { useQuery } from "@tanstack/react-query"
 import { useSession } from "next-auth/react"
 
 interface Params {
@@ -11,7 +11,7 @@ export const useGetUserLogs = ({ calMonth }: Params) => {
 
   const year = Number(calMonth.split(".")[0])
   const month = Number(calMonth.split(".")[1])
-  return useSuspenseQuery({
+  return useQuery({
     queryKey: ["userLogs", data?.user?.id, calMonth],
     queryFn: () => getUserLogs(year, month),
   })
